# examples/server/config-variations/same-package-multiple-files/api/types.go
// Code generated by oapi-codegen. DO NOT EDIT.

package api

import (
	"github.com/doordash-oss/oapi-codegen-dd/v3/pkg/runtime"
)

type HealthStatus struct {
	Status string `json:"status" validate:"required"`
}

func (h HealthStatus) Validate() error {
	return runtime.ConvertValidatorError(typesValidator.Struct(h))
}

type User struct {
	ID    string `json:"id" validate:"required"`
	Name  string `json:"name" validate:"required"`
	Email string `json:"email" validate:"required"`
}

func (u User) Validate() error {
	return runtime.ConvertValidatorError(typesValidator.Struct(u))
}

type CreateUserRequest struct {
	Name  string `json:"name" validate:"required"`
	Email string `json:"email" validate:"required"`
}

func (c CreateUserRequest) Validate() error {
	return runtime.ConvertValidatorError(typesValidator.Struct(c))
}

type Error struct {
	Code    *string `json:"code,omitempty"`
	Message *string `json:"message,omitempty"`
}

func (s Error) Error() string {
	return "unmapped client error"
}